requirements.txt
PyYAML==6.0
tqdm==4.62.3

# PyTorch
# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==1.12.1+cu102
# torchvision==0.13.1+cu102

pandas==1.4.3

tifffile==2022.8.12

# from Pytorch_UNet
matplotlib
Pillow
tqdm
wandb
scikit-image
albumentations

jupyterlab
notebook